requirements.txt
django~=5.0
celery~=5.3.0
coverage~=7.6.10
drf-spectacular~=0.28.0
drf-spectacular-sidecar~=2024.12.1
requests~=2.31.0
ujson~=5.9.0
pytest-django~=4.9.0
django-jet-reboot~=1.3.10
pyyaml~=6.0.0
django-rest-passwordreset>=1.3.0
